Move rpm_util.py into utils/
*OSBS-8605

Also change files which import rpm_util.py
