+++ atomic_reactor/plugins/pre_flatpak_create_dockerfile.py
@@ -21,7 +21,7 @@ from osbs.repo_utils import ModuleSpec
 from atomic_reactor.constants import DOCKERFILE_FILENAME, RELATIVE_REPOS_PATH, YUM_REPOS_DIR
 from atomic_reactor.plugin import PreBuildPlugin
 from atomic_reactor.plugins.pre_reactor_config import get_flatpak_base_image
-from atomic_reactor.rpm_util import rpm_qf_args
+from atomic_reactor.utils.rpm import rpm_qf_args
 
 
 # /var/tmp/flatpak-build is the final image we'll turn into a Flaptak
