+++ atomic_reactor/utils/rpm.py
